Trying sticks, wood, and problems faced
Here's a new story, of how the house was built. Unfortunately, I'm not a
great story teller yet, so I have missed a small detail while planning
my chapters, so this references another story... On hold for now!
Yeah I know, this is boring but follow me, keep up! We're not actually
writting a story here, just checking how git is a story of the project
you've built.
The plot was kind of down, so I went for some Michael Bay action, but
with words... Let me know if that works!